+++ src/Command/CommandInterface.php
@@ -14,7 +14,7 @@ use Yiisoft\Db\Constant\IndexType;
 use Yiisoft\Db\Constant\PseudoType;
 use Yiisoft\Db\Constant\ReferentialAction;
 use Yiisoft\Db\Exception\Exception;
-use Yiisoft\Db\Exception\InvalidArgumentException;
+use InvalidArgumentException;
 use Yiisoft\Db\Exception\InvalidCallException;
 use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
@@ -617,7 +617,7 @@ interface CommandInterface
      *
      * Note: The method will quote the `table` and `columns` parameter before using it in the generated SQL.
      */
-    public function insertWithReturningPks(string $table, array|QueryInterface $columns): array|false;
+    public function insertReturningPks(string $table, array|QueryInterface $columns): array|false;
 
     /**
      * Prepares the SQL statement to be executed.